Safer file reading
Files are now being read in a process that waits when the file could
not be retrieved immediately.  This helps in the process where files
are quickly written, and then read.  Often, this yielded errors,
thereby making the mu-har-transformation crash.  The current version
is much safer in this regard, only crashing after ten successive
failures of two seconds of waiting.